requirements.txt
tqdm 
pillow 
imagehash 
pandas